Don't use in-place editing with sed
In place editing with sed is not covered by POSIX, and only supported by GNU
sed (typically found on Linux) and BSED sed (e.g. used on macOS). But the
semantics of the -i option differ between the two in an incompatible way.

Instead of relying on non-portable behaviour, just put the sed output into a
temporary file.